adjusting button logic for login/logout and add a feeling

+++ src/containers/Main.js
@@ -12,12 +12,14 @@ class Main extends React.Component {
       user: localStorage.getItem("jwt")
     }
     this.fetchPosts = this.fetchPosts.bind(this)
-    this.addFeeling = this.addFeeling.bind(this)
-    this.logOut = this.logOut.bind(this)
   }
 
   componentWillMount() {
     this.fetchPosts();
+  }
+
+  componentDidUpdate() {
+    this.actionButtonText();
   }
 
   fetchPosts() {
@@ -30,15 +32,34 @@ class Main extends React.Component {
     })
   }
 
-  addFeeling() {
-    window.location.href = 'http://localhost:3001/newpost'
+  feelingButton() {
+    if (this.state.user === null) {
+      window.location.href = 'http://localhost:3001/login'
+    } else {
+      window.location.href = 'http://localhost:3001/newpost'
+    }
   }
 
-  logOut(ev) {
-    ev.preventDefault()
-    console.log("Logged OUTTTT")
-    localStorage.clear();
-    window.location.href = 'http://localhost:3001/login'
+  actionButtonText() {
+    const button = document.getElementById("user-register")
+    const feelbutton = document.getElementById('feel-button')
+    if (this.state.user === null) {
+      button.textContent = "Sign In"
+      feelbutton.textContent = "Login to Add a Feeling"
+    } else {
+      button.textContent = "Logout"
+      feelbutton.textContent = "Add a Feeling"
+
+    }
+  }
+
+  actionButtonFunction() {
+    if (this.state.user === null) {
+      window.location.href = 'http://localhost:3001/login'
+    } else {
+      localStorage.clear();
+      window.location.href = 'http://localhost:3001/login'
+    }
   }
 
   render() {
@@ -46,11 +67,11 @@ class Main extends React.Component {
       return <div />
     }
     return (
-      <div>
-        <button className="button" onClick={(ev)=>this.logOut(ev)}>Logout</button>
+      <div id="main-div">
+        <button id="user-register" className="button" onClick={()=>this.actionButtonFunction()}>Logout</button>
         <p>Current User ID: {this.state.user}</p>
         <div id="add-post-button">
-          <button className="button" type="button" onClick={() => this.addFeeling()}>Add Your Feelings</button>
+          <button id="feel-button" className="button" type="button" onClick={() => this.feelingButton()}>Add Your Feelings</button>
         </div>
         <Post posts={this.state.posts} />
       </div>
